Allow passing in celery app/modify timestamp
The current implementation of the package won't actually grab the
correct celery app. We're going to add the ability to pass in a celery
app. We're also modifying how we get the timestamp to pass to cloudwatch
because the timestamp from datetime.utcnow() is too far in the future
for cloudwatch to accept.

ticket: RM-900