Stream task-received to cloudwatch
Sometimes a task gets stuck in running mode for a long time, and while
we know the task exists, we don't see it in cloudwatch because it didn't
reach an end result. We're going to also stream received events to
cloudwatch so we know when a task makes it to a worker.

ticket: RM-900